Refactor `_converse.XMPPStatus` out into `headless/converse-status.js`
Also move some other methods out of `converse-core` into the plugins
that use them.
